no need to cast patch->columnofs to uint16_t

+++ i_astv16.c
@@ -958,7 +958,7 @@ void V_DrawPatchScaled(int16_t x, int16_t y, const patch_t __far* patch)
 		else if (dc_x >= SCREENWIDTH)
 			break;
 
-		const column_t *column = (const column_t *)((const byte *)patch + (uint16_t)patch->columnofs[col >> 8]);
+		const column_t *column = (const column_t *)((const byte *)patch + patch->columnofs[col >> 8]);
 
 		x = dc_x & 7;
 		uint32_t andmask = ~(0x80808080 >> x);
